Tansformed app to package and separeted each module
