+++ stockapp/static/main.js
