+++ stockapp/templates/base.html
