+++ stockapp/templates/drinks.html
