+++ stockapp/templates/index.html
