+++ stockapp/templates/list.html
