+++ stockapp/templates/login.html
