+++ stockapp/templates/register.html
