+++ stockapp/templates/table.html
